Move to an underscore-less package
