+++ android-junit5/src/test/groovy/de/mannodermaus/gradle/plugins/junit5/CallableGroovyTest.groovy
@@ -2,9 +2,9 @@ package de.mannodermaus.gradle.plugins.junit5
 
 import org.junit.Test
 
-class CallableFromGroovyTests {
+class CallableGroovyTest {
   @Test
-  void callableSyntaxWorks() {
+  void syntaxWorks() {
     def callable = new Callable({ 2 + 2 })
     assert callable() == 4
   }
