+++ android-junit5/src/test/groovy/de/mannodermaus/gradle/plugins/junit5/util/TestEnvironment.groovy
@@ -1,4 +1,4 @@
-package de.mannodermaus.gradle.plugins.android_junit5.util
+package de.mannodermaus.gradle.plugins.junit5.util
 
 import com.android.annotations.Nullable
