+++ android-junit5/src/test/kotlin/de/mannodermaus/gradle/plugins/junit5/CallableKotlinTest.kt
@@ -2,9 +2,9 @@ package de.mannodermaus.gradle.plugins.junit5
 
 import org.junit.Test
 
-class CallableFromKotlinTests {
+class CallableKotlinTest {
   @Test
-  fun callableSyntaxWorks() {
+  fun syntaxWorks() {
     val callable = Callable { 2 + 2 }
     assert(callable() == 4)
   }
